Add empty HelloWorld constructor with explanatory comment
Added an empty constructor `HelloWorld()` to the `HelloWorld` class in the `example_web_app.Classes` namespace. The constructor includes a comment explaining that it is not used in the example but demonstrates how to use a constructor. Also made some formatting changes, such as adding new lines. The `GetMessage` method and the `IHelloWorld` interface remain unchanged.

AB#114382